mean, std. dev in results screen & sprint feature
Added a "sprint" feature in gameplay that lets the player move 2 spaces at a time when they hold a shift key and then press an arrow key. Also added mean and standard deviation to the results screen and changed arrangement/size of some text. Also added a basic readme for the time being